+++ changelog.md
@@ -1,3 +1,10 @@
 # Changelog
 
-Demo was unaffected by module updates.
+**./index.js**
+* Declared new global variables:
+	* putComp
+	* delComp
+* Renamed global variables:
+	* 'getComplete' to 'getComp'
+	* 'postComplete' to 'postComp'
+* The URL in 'sendPostRequest' is now a global variable 'modifyURL'